# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_audit_audit_parse_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_audit_audit_parse_post",
        "type": "object"
      },
...
      "Envelope_dict_str__Any__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "mock": {
            "const": false,
            "default": false,
            "title": "Mock",
            "type": "boolean"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          }
        },
        "required": [
          "model_version",
          "data"
        ],
        "title": "Envelope[dict[str, Any]]",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "needs": {
            "items": {
              "type": "string"
            },
            "title": "Needs",
            "type": "array"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Early-warning and stress-test API over frozen, checksummed models. Every number carries the tool_result_id that produced it.",
    "title": "COOKED \u2014 Synthetic Data Only",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/audit/parse": {
      "post": {
        "description": "Sample audits map to their synthetic student; real PDFs are read by Gemini into a profile.\nThe file is never stored; only parsed courses and terms are kept.",
        "operationId": "audit_audit_parse_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_audit_audit_parse_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_str__Any__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "summary": "Audit",
        "tags": [
          "product"
        ]
      }
    },
...
  }
}